requirements.txt
requests~=2.28.1
kafka-python